Added a missing validation
It was possible to have two transitions starting from the same state
triggered by the same trigger.

This would lead to a non-deterministic finite automata and since we
don't convert it to a DFA this would either cause to undefined behavior
or a silently ignored second transition.

Now this is checked during validation and rejected.